backend/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
python-multipart
email-validator
sentence-transformers
scikit-learn
numpy
groq
python-dotenv
